requirement.txt
pandas
beautifulsoup4
newspaper3k
requests
textblob
textstat
nltk
openpyxl
